feat(front): improve settings
