+++ src/lib/components/settings/Preferences.svelte
@@ -1,0 +1,20 @@
+<script lang="ts">
+	import { InlineSelect, Switch } from '~/lib/components';
+	import { settings } from '~/lib/stores';
+	import type { Settings } from '~/lib/types';
+
+	const options: Array<Settings['notificationAxis']> = ['Auto', 'Vertical', 'Horizontal'];
+</script>
+
+<div class="wrapper">
+	<Switch
+		label="Activate push notifications (only on desktop app)"
+		bind:active={$settings.activateNotifications}
+	/>
+	<Switch
+		label="Mark an event as read when opening in the browser"
+		bind:active={$settings.readWhenOpenInBrowser}
+	/>
+	<Switch label="Mark an event as read when pinned" bind:active={$settings.readWhenPin} />
+	<InlineSelect label="Notification axis" {options} bind:value={$settings.notificationAxis} />
+</div>
